+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - New `--merge` flag merges multiple `output.lammpstrj*` files into one.
 - New `tests` directory for testing with pytest and tox.
+- New `docs` directory for sphinx documentation.
 
 ### Fixed